timer2 to do housekeeping

+++ main.h
@@ -22,6 +22,6 @@ void main_init(void);
 #define  FAT_YEAR_MASK          0x7F
 
 uint8_t  validDate(uint8_t day, uint8_t month, uint8_t year);
-extern volatile uint8_t timer2_sleep_delay;
+extern volatile uint16_t timer2_sleep_delay;
 
 #endif
